requirements.txt
# JurDatasetBrasil - Dependências Python
# Python 3.10+

# Processamento de documentos
docling>=2.0.0
pypdf2>=3.0.0
python-docx>=1.1.0

# LLMs e APIs
openai>=1.0.0
langchain>=0.1.0
anthropic>=0.21.0
requests>=2.31.0

# Supabase e Database
supabase>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0

# Embeddings e NLP
sentence-transformers>=2.2.0
tiktoken>=0.5.0
nltk>=3.8.1

# Utilitários
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0
jsonlines>=4.0.0
pyyaml>=6.0.1

# Validação e qualidade
jsonschema>=4.20.0
pandas>=2.1.0
numpy>=1.24.0

# Logging e monitoramento
loguru>=0.7.0

# Desenvolvimento (opcional)
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0

# Dashboard
streamlit>=1.31.0
plotly>=5.18.0
watchdog>=4.0.0


# API
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv
psycopg2
